Normalizing components and other refactor
