+++ .geminiignore
@@ -1,0 +1,2 @@
+*.env
+local_testing/
