+++ frontend/src/components/common/PRStub.vue
