+++ frontend/src/components/layout/DashboardPeek.vue
