+++ frontend/src/components/layout/FAQSection.vue
