+++ frontend/src/components/layout/FeaturesGrid.vue
