+++ frontend/src/components/layout/HeroSection.vue
