+++ frontend/src/components/layout/PricingSection.vue
