+++ frontend/src/components/layout/ProblemImpact.vue
