+++ frontend/src/components/layout/RealTimeCounter.vue
