+++ frontend/src/components/layout/TransparencySteps.vue
