+++ frontend/src/components/view/auth/DonorLoginForm.vue
@@ -63,7 +63,7 @@
   </div>
 </template>
 
-<script setup>
+<script setup lang="ts">
 import { ref } from 'vue';
 
 const isValid = ref(false);
