+++ frontend/src/components/view/auth/RegisterForm.vue
@@ -80,7 +80,7 @@
   </div>
 </template>
 
-<script setup>
+<script setup lang="ts">
 import { ref } from 'vue';
 
 const isValid = ref(false);
