+++ frontend/src/views/HomeView.vue
@@ -12,14 +12,14 @@
 </template>
 
 <script setup lang="ts">
-import HeroSection from '@/components/landing/HeroSection.vue';
-import ProblemSolution from '@/components/landing/ProblemSolution.vue';
-import FeaturesGrid from '@/components/landing/FeaturesGrid.vue';
-import UseCases from '@/components/landing/UseCases.vue';
-import DashboardPeek from '@/components/landing/DashboardPeek.vue';
-import Testimonials from '@/components/landing/Testimonials.vue';
-import PricingSection from '@/components/landing/PricingSection.vue';
-import FinalCTA from '@/components/landing/FinalCTA.vue';
+import HeroSection from '@/components/layout/HeroSection.vue';
+import ProblemSolution from '@/components/layout/ProblemSolution.vue';
+import FeaturesGrid from '@/components/layout/FeaturesGrid.vue';
+import UseCases from '@/components/layout/UseCases.vue';
+import DashboardPeek from '@/components/layout/DashboardPeek.vue';
+import Testimonials from '@/components/layout/Testimonials.vue';
+import PricingSection from '@/components/layout/PricingSection.vue';
+import FinalCTA from '@/components/layout/FinalCTA.vue';
 import type { HomePageConfig } from '@/config/tenantConfig';
 
 defineProps<{
